+++ src/imio/events/core/upgrades.py
@@ -16,3 +16,10 @@ def refresh_objects_faceted(context):
         obj = brain.getObject()
         reload_faceted_config(obj, request)
         logger.info("Faceted refreshed on {}".format(obj.Title()))
+
+
+def add_event_dates_index(context):
+    catalog = api.portal.get_tool("portal_catalog")
+    catalog.addIndex("event_dates", "KeywordIndex")
+    catalog.manage_reindexIndex(ids=["event_dates"])
+    logger.info("Added and indexed event_dates KeywordIndex")